Add label to search button
When you're making a product to be consumed by english speakers it's ok
to have a hard coded "Search" label to your search button. But if you
want to attend a different language, such as brazilian portuguese or
japanese, this strategy will no longer work.

Facing that problem, I create a solution that I think is good not only
for me but for everyone using Rummage and is good to Rummage itself
because this adds more flexibility to use, helping to attract a bigger
range of users.

The change itself is not a big deal, it just pass around an label
attribute that is used as substitute of the hard coded "Search" label.
Works identical as the button class attribute which is passed in the
search_form params.