allow a rummage_key to be set on sort
the sort helper used to only allow 'rummage' as the key. On a page with multiple rummage elements you need to specify the rummage key for each. You can pass this in as an option. I also added a more consise systax, where the second element can be a tuple with the rummage schema and the rummage_key.

This allows for a more consise syntax (than pass)